factorio/requirements.txt
ollama>=0.1.0
factorio-rcon-py>=0.1.0
requests>=2.31.0